+++ src/components/shared/index.js
@@ -1,5 +1,8 @@
 import RegistrationNavbar from './RegistrationNavbar';
 import Button from './Button';
 import NavigationButton from './NavigationButton';
+import LeftBigImage from './LeftBigImage';
 
-export { RegistrationNavbar, Button, NavigationButton };
+export {
+  RegistrationNavbar, Button, NavigationButton, LeftBigImage,
+};
