+++ src/pages/ReservationsCreate.js
@@ -1,18 +1,16 @@
 import { useEffect, useState } from 'react';
-import {
-  Box, HStack, Text, Flex, Image,
-} from '@chakra-ui/react';
-import ReactDatePicker from 'react-datepicker';
-import addDays from 'date-fns/addDays';
 import { useParams, useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
+import {
+  HStack, Text, Flex,
+} from '@chakra-ui/react';
 import { format as formatDate } from 'date-fns';
 import Layout from '../components/Layout/Layout';
 import { createApi } from '../utils/helpers';
 import { useAuth } from '../utils/customHooks';
 import { fetchShifts } from '../utils/actions/shifts.actions';
-import { NavigationButton } from '../components/shared';
-import DatePicker from '../components/ReservationsCreate/DatePicker';
+import { LeftBigImage } from '../components/shared';
+import Header from '../components/ReservationsCreate/Header';
 
 const ReservationsCreate = () => {
   useAuth('/restaurants', ['', 'admin']);
@@ -76,42 +74,15 @@ const ReservationsCreate = () => {
   return (
     <Layout>
       <HStack h="100vh" spacing={0} position="relative">
-        <Box h="100vh" flex="1 1 65%">
-          <Image src="../../assets/images/details.jpg" alt="" objectFit="cover" h="full" w="full" />
-        </Box>
-        <NavigationButton position="absolute" bottom={40} onClick={() => { navigate(-1); }} isReversed />
+        <LeftBigImage src="../../assets/images/details.jpg" />
         <Flex
           h="full"
           direction="column"
           flex="1 0 600px"
           alignItems="center"
-          pt={24}
-          px={24}
           overflow="auto"
-          bg="green.300"
-          color="white"
         >
-          <Box px={8} w="full" mb={8}>
-            <HStack
-              minW="full"
-              bg="white"
-              rounded="full"
-              position="relative"
-              minH={16}
-              justify="center"
-            >
-              <Box position="absolute" right={0} top={0} bottom={0}>
-                <DatePicker date={date} setDate={setDate} />
-              </Box>
-              <Text
-                color="green.300"
-                fontSize="xl"
-                fontWeight="BOLD"
-              >
-                {formatDate(date, 'do MMMM')}
-              </Text>
-            </HStack>
-          </Box>
+          <Header date={date} setDate={setDate} />
           <Text as="h4" fontSize="lg" w="full" mb={4}>Shifts available:</Text>
           {error && (
           <Text>
@@ -133,7 +104,3 @@ const ReservationsCreate = () => {
 };
 
 export default ReservationsCreate;
-
-// <Box mb={4}>
-//   <DateStripe date={date} setDate={setDate} />
-// </Box>;
