+++ src/pages/RestaurantDetails.js
@@ -1,20 +1,19 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useParams, useNavigate } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import {
-  HStack, Image, Box, Text, Flex, Badge, Grid, Icon,
+  HStack, Text, Flex, Badge, Grid, Icon,
 } from '@chakra-ui/react';
 import { IoBookOutline, IoLogoUsd } from 'react-icons/io5';
 import Layout from '../components/Layout/Layout';
 import { fetchSingleRestaurant } from '../utils/actions/restaurants.actions';
 import { useAuth } from '../utils/customHooks';
-import { NavigationButton, Button } from '../components/shared';
+import { Button, LeftBigImage } from '../components/shared';
 import MenuImage from '../components/Restaurants/DetailsPage/MenuImage';
 
 const RestaurantDetails = () => {
   useAuth('/restaurants', ['', 'admin']);
   const dispatch = useDispatch();
-  const navigate = useNavigate();
 
   let { id } = useParams();
 
@@ -35,10 +34,7 @@ const RestaurantDetails = () => {
     <Layout>
       {selectedRestaurant && (
         <HStack h="100vh" spacing={0} position="relative">
-          <Box h="100vh" flex="1 1 65%">
-            <Image src="../assets/images/details.jpg" alt={name} objectFit="cover" h="full" w="full" />
-          </Box>
-          <NavigationButton position="absolute" bottom={40} onClick={() => { navigate(-1); }} isReversed />
+          <LeftBigImage src="../assets/images/details.jpg" />
           <Flex
             h="full"
             direction="column"
